test(SENTZ-5683): Split delegate-lookup cases out of CertificateTests
CertificateTests sits near the file-length warning, and every case that
drives the pinning delegate's challenge handling is one coherent group.
Moving that group into its own file keeps both files well under the limit.

The alphaNet intermediate lookup and the delegate accessor move to file
scope so both test files can share them without duplicating the base64
lookup logic.
