+++ Tests/Unit/Network/Certificates/CertificateTests.swift
@@ -12,10 +12,16 @@ import XCTest
 
 // The hosts a test pins against. `pinned` covers a case where one host is
 // enough, and `fog` and `consensus` separate the two setters' roots.
-private enum TestHost {
+enum TestHost {
     static let pinned = "example.com"
     static let fog = "fog.example.com"
     static let consensus = "consensus.example.com"
+}
+
+func pinningDelegate(
+    of requester: DefaultHttpRequester
+) throws -> CertificatePinningDelegate {
+    try XCTUnwrap(requester.session.delegate as? CertificatePinningDelegate)
 }
 
 class CertificateTests: XCTestCase {
@@ -232,7 +238,7 @@ class CertificateTests: XCTestCase {
         config.consensusTrustRoots[.http] = nil
         config.fogTrustRoots[.http] = nil
         let requester = DefaultHttpRequester()
-        let roots = try alphaNetCertificates(.valid)
+        let roots = try SecCertificateTests.Fixtures.AlphaNet.certificates(.valid)
         requester.setFogTrustRoots(roots, hosts: [TestHost.fog])
 
         config.httpRequester = requester
@@ -290,141 +296,6 @@ class CertificateTests: XCTestCase {
             pinned.publicKeys)
     }
 
-    // The cases below drive both delegate shims across `handle`'s outcomes.
-
-    func testServerTrustMatchingAPinnedKeyIsAccepted() throws {
-        let fixture = try SecCertificateTests.Fixtures.AlphaNet()
-        let requester = DefaultHttpRequester()
-        requester.setFogTrustRoots(try alphaNetCertificates(.valid), hosts: [TestHost.pinned])
-
-        let result = try answer(of: requester, against: fixture.secTrust)
-
-        XCTAssertEqual(result.disposition, .useCredential)
-        XCTAssertNotNil(result.credential)
-    }
-
-    // The consensus roots pin this host and don't carry the fixture chain, so
-    // the challenge will be refused.
-    func testServerTrustMatchingNoPinnedKeyIsCancelled() throws {
-        let fixture = try SecCertificateTests.Fixtures.AlphaNet()
-        let requester = DefaultHttpRequester()
-        requester.setConsensusTrustRoots(try alphaNetCertificates(.wrong), hosts: [TestHost.pinned])
-
-        XCTAssertEqual(
-            try answer(of: requester, against: fixture.secTrust).disposition,
-            .cancelAuthenticationChallenge)
-    }
-
-    // The fog roots carry this fixture chain and the consensus roots don't, so
-    // the consensus host will be refused while the fog host is accepted.
-    func testTheRootsOfOneHostDoNotPinAnother() throws {
-        let fixture = try SecCertificateTests.Fixtures.AlphaNet()
-        let requester = DefaultHttpRequester()
-        requester.setFogTrustRoots(try alphaNetCertificates(.valid), hosts: [TestHost.fog])
-        requester.setConsensusTrustRoots(
-            try alphaNetCertificates(.wrong), hosts: [TestHost.consensus])
-
-        let refused = try answer(
-            of: requester, against: fixture.secTrust, host: TestHost.consensus)
-        XCTAssertEqual(refused.disposition, .cancelAuthenticationChallenge)
-        let accepted = try answer(of: requester, against: fixture.secTrust, host: TestHost.fog)
-        XCTAssertEqual(accepted.disposition, .useCredential)
-    }
-
-    // Neither setter names the challenged host, so every pinned root takes part
-    // in its challenge.
-    func testAHostNoSetterNamedIsJudgedAgainstEveryPinnedRoot() throws {
-        let fixture = try SecCertificateTests.Fixtures.AlphaNet()
-        let requester = DefaultHttpRequester()
-        requester.setFogTrustRoots(try alphaNetCertificates(.valid), hosts: [TestHost.fog])
-        let judged = try answer(of: requester, against: fixture.secTrust, host: TestHost.consensus)
-        XCTAssertEqual(judged.disposition, .useCredential)
-    }
-
-    // With no roots set there is nothing to pin against, so the challenge goes
-    // to the system rather than being refused.
-    func testServerTrustWithoutPinnedKeysFallsThroughToDefaultHandling() throws {
-        let fixture = try SecCertificateTests.Fixtures.AlphaNet()
-
-        XCTAssertEqual(
-            try answer(of: DefaultHttpRequester(), against: fixture.secTrust).disposition,
-            .performDefaultHandling)
-    }
-
-    func testChallengeWithoutServerTrustIsCancelled() throws {
-        XCTAssertEqual(
-            try answer(of: DefaultHttpRequester(), against: nil).disposition,
-            .cancelAuthenticationChallenge)
-    }
-
-    // A URLSession holds its delegate until it is invalidated, so a requester
-    // that goes out of scope without invalidating leaves its delegate behind.
-    func testRequesterReleasesItsDelegateWhenItGoesOutOfScope() {
-        weak var delegate: CertificatePinningDelegate?
-        autoreleasepool {
-            let requester = DefaultHttpRequester()
-            delegate = requester.session.delegate as? CertificatePinningDelegate
-            XCTAssertNotNil(delegate)
-        }
-
-        // Invalidation is asynchronous, so the release lands on the session's own
-        // queue rather than on this one.
-        let deadline = Date().addingTimeInterval(5)
-        while delegate != nil && Date() < deadline {
-            RunLoop.current.run(until: Date().addingTimeInterval(0.05))
-        }
-
-        XCTAssertNil(delegate)
-    }
-
-    // Distinct fog and consensus roots let each lookup below name the set or
-    // sets that answered it.
-    func testTheSetThatNamesAHostAnswersForIt() throws {
-        let requester = DefaultHttpRequester()
-        let delegate = try pinningDelegate(of: requester)
-        let fog = try alphaNetCertificates(.valid)
-        let consensus = try alphaNetCertificates(.wrong)
-        XCTAssertNotEqual(fog.publicKeys, consensus.publicKeys)
-
-        requester.setFogTrustRoots(fog, hosts: [TestHost.fog.uppercased()])
-        XCTAssertEqual(delegate.pinnedKeys(for: TestHost.consensus), fog.publicKeys)
-
-        requester.setConsensusTrustRoots(consensus, hosts: [TestHost.consensus + "."])
-        XCTAssertEqual(delegate.pinnedKeys(for: TestHost.fog), fog.publicKeys)
-        XCTAssertEqual(delegate.pinnedKeys(for: TestHost.consensus + ".."), consensus.publicKeys)
-
-        let noKeys = try XCTUnwrap(SecSSLCertificates(trustRootBytes: []))
-        requester.setFogTrustRoots(noKeys, hosts: [TestHost.fog])
-        XCTAssertEqual(delegate.pinnedKeys(for: TestHost.fog), consensus.publicKeys)
-
-        requester.setFogTrustRoots(nil, hosts: [])
-        XCTAssertEqual(delegate.pinnedKeys(for: TestHost.fog), consensus.publicKeys)
-
-        requester.setFogTrustRoots(fog, hosts: [TestHost.consensus])
-        XCTAssertEqual(
-            delegate.pinnedKeys(for: TestHost.consensus),
-            fog.publicKeys + consensus.publicKeys)
-        XCTAssertEqual(
-            delegate.pinnedKeys(for: TestHost.pinned),
-            fog.publicKeys + consensus.publicKeys)
-    }
-
-    private enum AlphaNetIntermediate { case valid, wrong }
-
-    private func alphaNetCertificates(
-        _ intermediate: AlphaNetIntermediate
-    ) throws -> SecSSLCertificates {
-        let base64: String
-        switch intermediate {
-        case .valid:
-            base64 = SecCertificateTests.Fixtures.AlphaNet.intermediateCertificateBase64
-        case .wrong:
-            base64 = SecCertificateTests.Fixtures.AlphaNet.wrongIntermediateCertificateBase64
-        }
-        let bytes = try XCTUnwrap(Data(base64Encoded: base64))
-        return try XCTUnwrap(try SecSSLCertificates(trustRootBytes: [bytes]))
-    }
-
     // `validateAgainst` calls back on the calling thread, so the reason is set
     // before this returns.
     private func refusal(of trust: SecTrust, against pinnedKeys: [SecKey]) throws -> String {
@@ -435,57 +306,5 @@ class CertificateTests: XCTestCase {
             }
         }
         return try XCTUnwrap(reason)
-    }
-
-    private func pinningDelegate(
-        of requester: DefaultHttpRequester
-    ) throws -> CertificatePinningDelegate {
-        try XCTUnwrap(requester.session.delegate as? CertificatePinningDelegate)
-    }
-
-    // Server trust is a session-level challenge. `validateAgainst` calls back on
-    // the calling thread, so both shims answer before this returns.
-    private func answer(
-        of requester: DefaultHttpRequester,
-        against trust: SecTrust?,
-        host: String = TestHost.pinned
-    ) throws -> (disposition: URLSession.AuthChallengeDisposition?, credential: URLCredential?) {
-        let space: URLProtectionSpace
-        if let trust = trust {
-            space = TrustingProtectionSpace(trust: trust, host: host)
-        } else {
-            space = URLProtectionSpace(
-                host: host,
-                port: 443,
-                protocol: NSURLProtectionSpaceHTTPS,
-                realm: nil,
-                authenticationMethod: NSURLAuthenticationMethodServerTrust)
-        }
-        let challenge = URLAuthenticationChallenge(
-            protectionSpace: space,
-            proposedCredential: nil,
-            previousFailureCount: 0,
-            failureResponse: nil,
-            error: nil,
-            sender: NullChallengeSender())
-
-        let delegate = try pinningDelegate(of: requester)
-        let url = try XCTUnwrap(URL(string: "https://example.com"))
-        let task = requester.session.dataTask(with: url)
-
-        var sessionDisposition: URLSession.AuthChallengeDisposition?
-        var credential: URLCredential?
-        delegate.urlSession(requester.session, didReceive: challenge) { result, resultCredential in
-            sessionDisposition = result
-            credential = resultCredential
-        }
-
-        var taskDisposition: URLSession.AuthChallengeDisposition?
-        delegate.urlSession(requester.session, task: task, didReceive: challenge) { result, _ in
-            taskDisposition = result
-        }
-
-        XCTAssertEqual(sessionDisposition, taskDisposition)
-        return (sessionDisposition, credential)
     }
 }
